added correct styling on title and refactored into folders
