+++ src/index.js
@@ -1,7 +1,6 @@
 export { default as CheckInfoCard } from './CheckInfoCard';
 export { default as TranslationHelps } from './TranslationHelps';
 export { default as ScripturePane } from './ScripturePane';
-//export { default as ExpandedHelpsModal } from './ExpandedHelpsModal';
 export { default as VerseEditor } from './VerseEditor';
 export { default as VerseCheck } from './VerseCheck';
 export { default as GroupMenu } from './GroupMenu';
